Extract template-engine utilities into workspace package
