+++ Work/package.json
@@ -62,6 +62,7 @@
   "license": "MIT",
   "private": false,
   "workspaces": [
+    "../packages/template-engine",
     "../sub-modules/Typography",
     "../sub-modules/innerComponents",
     "../sub-modules/Miscellaneous"
@@ -73,6 +74,7 @@
     "access": "public"
   },
   "dependencies": {
+    "@llazyemail/template-engine": "0.1.0",
     "atherdon-newsletter-js-layouts-body": "^3.2.0",
     "atherdon-newsletter-js-layouts-misc": "^3.0.0",
     "atherdon-newsletter-js-layouts-typography": "^3.1.0",
