+++ packages/template-engine/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@llazyemail/template-engine",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Shared template engine utilities for registry, definitions, and validation.",
+  "main": "src/index.js",
+  "license": "MIT"
+}
