+++ packages/template-engine/src/index.js
@@ -1,0 +1,6 @@
+export { createTemplateFromDefinition } from './createTemplateFromDefinition';
+export {
+  validateTemplateInput,
+  validateHnTemplateInput,
+  validateHnWithoutAdsTemplateInput,
+} from './validation';
